refactor: separate symmetric and asymmetric encoders
